+++ R/metainc.R
@@ -665,7 +665,7 @@ metainc <- function(event.e, time.e, event.c, time.c, studlab,
                     text.random = gs("text.random"),
                     text.predict = gs("text.predict"),
                     text.w.fixed = gs("text.w.fixed"),
-                    text.w.random = gs("test.w.random"),
+                    text.w.random = gs("text.w.random"),
                     ##
                     title = gs("title"), complab = gs("complab"),
                     outclab = "",